+++ frontend/src/app/login/page.tsx
@@ -13,7 +13,7 @@ export default function LoginPage() {
   const [error, setError] = useState("");
   const [isLoading, setIsLoading] = useState(false);
   const handleBackToSelection = () => {
-    router.push("/login");
+    router.push("/login/type");
   };
   const handleChange = (e: React.ChangeEvent<HTMLInputElement>) => {
     const { name, value } = e.target;
